requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
python-multipart>=0.0.6
python-decouple>=3.8
pydantic>=2.6.0
pydantic-settings>=2.1.0
exa-py>=1.0.0
python-dotenv>=1.0.0

# AI Pipeline Dependencies
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
openai>=1.10.0

# PDF Processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
pytesseract>=0.3.10
Pillow>=10.0.0

# Background Tasks
celery>=5.3.0
redis>=5.0.0

# Scientific APIs
requests>=2.31.0
httpx>=0.27.0